refactor(chains): hoist + generalize vault state machine to chains::vault (Monad behavior unchanged)
Move the foreign-chain vault record, error types, CR math, and the
open/verify/withdraw/close state machine out of chains::monad::chain_vault
into a new chain-agnostic chains::vault so Solana (Task 6) can reuse it.

- chains::vault holds the generalized helpers. ChainVaultStatus, ChainVaultV1
  (with the byte-identical #[serde(rename = "collateral_amount_e18")]),
  OpenVaultError, WithdrawError, collateral_ratio_e4, and
  verify_deposit_and_enqueue_mint_in_state move verbatim. The three
  address/price-aware helpers (open/withdraw/close) gain two parameters:
  address_validator: fn(&str) -> bool and price_symbol: &str, replacing the
  hardcoded is_valid_evm_address / "MON".
- chains::monad::chain_vault now re-exports the types + un-parameterized helper
  and wraps the three with their ORIGINAL signatures, baking in
  (is_valid_evm_address, "MON"). Every existing Monad caller and test resolves
  and behaves byte-identically; the candid interface is unchanged.
- Add SOLANA_MIN_CR_E4 (130%, matching Monad) to solana::config.
- Add chains::tests_vault exercising the new seam directly with an injected
  validator and the "SOL" price symbol.

Verification: full lib suite 318 passed; candid compat passes WITHOUT regen;
Monad PocketIC suites pass (phase1b_monad_happy_path, phase1c_burn_proof,
multi_chain_supply_invariant, phase1b_supply_gate, phase1b_burn_idempotency);
wasm release builds; serde unchanged at 1.0.217.
