+++ src/rumi_protocol_backend/src/chains/mod.rs
@@ -17,6 +17,7 @@ pub mod config;
 pub mod multi_chain_state;
 pub mod settlement_queue;
 pub mod supply;
+pub mod vault;
 pub mod monad;
 pub mod solana;
 
@@ -49,3 +50,6 @@ mod tests_admin;
 
 #[cfg(test)]
 mod tests_self_check;
+
+#[cfg(test)]
+mod tests_vault;
